+++ processing/src/main/java/org/apache/druid/query/aggregation/first/StringFirstAggregator.java
@@ -20,6 +20,7 @@
 package org.apache.druid.query.aggregation.first;
 
 import org.apache.druid.java.util.common.DateTimes;
+import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.query.aggregation.Aggregator;
 import org.apache.druid.query.aggregation.SerializablePairLongString;
 import org.apache.druid.segment.BaseLongColumnValueSelector;
@@ -54,7 +55,7 @@ public class StringFirstAggregator implements Aggregator
   @Override
   public void aggregate()
   {
-    final SerializablePairLongString inPair = StringAggregatorUtils.readPairFromSelectors(
+    final SerializablePairLongString inPair = StringFirstLastUtils.readPairFromSelectors(
         timeSelector,
         valueSelector
     );
@@ -72,7 +73,7 @@ public class StringFirstAggregator implements Aggregator
   @Override
   public Object get()
   {
-    return new SerializablePairLongString(firstTime, StringAggregatorUtils.chop(firstValue, maxStringBytes));
+    return new SerializablePairLongString(firstTime, StringUtils.chop(firstValue, maxStringBytes));
   }
 
   @Override
