+++ processing/src/main/java/org/apache/druid/query/aggregation/first/StringFirstLastUtils.java
@@ -29,22 +29,9 @@ import javax.annotation.Nullable;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 
-public class StringAggregatorUtils
+public class StringFirstLastUtils
 {
   private static final int NULL_VALUE = -1;
-
-  @Nullable
-  public static String chop(@Nullable final String s, final int maxBytes)
-  {
-    if (s == null) {
-      return null;
-    } else {
-      // Shorten firstValue to what could fit in maxBytes as UTF-8.
-      final byte[] bytes = new byte[maxBytes];
-      final int len = StringUtils.toUtf8WithLimit(s, ByteBuffer.wrap(bytes));
-      return new String(bytes, 0, len, StandardCharsets.UTF_8);
-    }
-  }
 
   @Nullable
   public static SerializablePairLongString readPairFromSelectors(
