+++ processing/src/main/java/org/apache/druid/query/aggregation/last/StringLastAggregator.java
@@ -20,9 +20,10 @@
 package org.apache.druid.query.aggregation.last;
 
 import org.apache.druid.java.util.common.DateTimes;
+import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.query.aggregation.Aggregator;
 import org.apache.druid.query.aggregation.SerializablePairLongString;
-import org.apache.druid.query.aggregation.first.StringAggregatorUtils;
+import org.apache.druid.query.aggregation.first.StringFirstLastUtils;
 import org.apache.druid.segment.BaseLongColumnValueSelector;
 import org.apache.druid.segment.BaseObjectColumnValueSelector;
 
@@ -52,7 +53,7 @@ public class StringLastAggregator implements Aggregator
   @Override
   public void aggregate()
   {
-    final SerializablePairLongString inPair = StringAggregatorUtils.readPairFromSelectors(
+    final SerializablePairLongString inPair = StringFirstLastUtils.readPairFromSelectors(
         timeSelector,
         valueSelector
     );
@@ -75,7 +76,7 @@ public class StringLastAggregator implements Aggregator
   @Override
   public Object get()
   {
-    return new SerializablePairLongString(lastTime, StringAggregatorUtils.chop(lastValue, maxStringBytes));
+    return new SerializablePairLongString(lastTime, StringUtils.chop(lastValue, maxStringBytes));
   }
 
   @Override
